feat(#167): add memo hover ui and split memo component
